+++ CHANGES.md
@@ -1,5 +1,7 @@
 0.5.0
 ==================
+* Added support for the PHP editor.
+* Added the Oblivion theme.
 
 0.4.0 (2011-01-13)
 ==================