+++ CHANGES.md
@@ -4,6 +4,7 @@
 * Added support for the Ant editor.
 * Added support for the SQL editor.
 * Added the Oblivion theme.
+* Moved preference page to General->Appereance.
 
 0.4.0 (2011-01-13)
 ==================